# apm/documentation/APM_API_OPENAPI.yaml
openapi: 3.0.0
info:
  title: APM (Approvals Management) API
  description: |
    REST API for the Approvals Management module. Used by approver apps to list pending approvals,
    view documents, perform actions (approve/reject/return/cancel), and access memo lists.
    **Other memos** (catalogue types under *Other memos* in the web app) appear in pending approvals as category **Other Memo**, document type path **other_memo**, and `item_type` **OtherMemo**.
    Authentication is JWT (Bearer token). API users are synced from the staff app `user` table;
    password verification supports Argon2i hashes.
  version: 1.0.0
  contact:
    name: APM Support
servers:
  - url: "{{API_BASE_URL}}"
    description: API base URL for this server (injected from the host where the docs are loaded)
  - url: "{{APP_BASE_URL}}"
    description: Web app root (document verification upload, memo print PDFs; use for paths under /api/documents/verify and print paths below)
...
paths:
...
  /pending-approvals:
    get:
      tags: [Pending]
      summary: Pending approvals (memos with approval trails)
      description: |
        List of documents pending the current user's action, grouped by category (Matrix, Special Memo, **Other Memo**, etc.).
        **Other Memo** items use `item_type` **OtherMemo** and a configurable approver chain (not workflow definitions).
        Each item includes **approval_trails**: full history, sorted by **created_at** descending (most recent first).
        Use this to show pending memos with their approval trail; use GET /documents/other_memo/{id} for full other-memo detail.
      operationId: pendingApprovals
      security: [{ bearerAuth: [] }]
      parameters:
        - name: category
          in: query
          schema: { type: string, example: all }
          description: Filter by category (e.g. Matrix, Special Memo) or "all"
      responses:
        '200':
          description: Pending items (each with approval_trails) and summary
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: object
                    properties:
                      pending:
                        type: object
                        description: Keys are category names (e.g. "Special Memo", "Matrix", "Other Memo"); values are arrays of pending items, each with approval_trails.
                        additionalProperties:
                          type: array
                          items: { $ref: '#/components/schemas/PendingApprovalItem' }
                      summary: { type: object, description: total_pending, by_category, by_division, oldest_pending, newest_pending }
                      filters: { type: object }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT from login or refresh
  schemas:
...
    ApprovalTrailItem:
      type: object
      properties:
        id: { type: integer }
        action: { type: string }
        remarks: { type: string, nullable: true }
        approval_order: { type: integer, nullable: true }
        staff_id: { type: integer, nullable: true }
        staff_name: { type: string, nullable: true }
        staff_image_url:
          type: string
          format: uri
          nullable: true
          description: Absolute URL for the acting staff member's portrait (APM profile photo or legacy staff upload).
        oic_staff_id: { type: integer, nullable: true }
        oic_staff_name: { type: string, nullable: true }
        oic_staff_image_url:
          type: string
          format: uri
          nullable: true
          description: Absolute URL for the OIC staff portrait when an officer-in-charge acted on behalf of the approver.
        role: { type: string, nullable: true }
        created_at: { type: string, format: date-time, nullable: true }
        is_archived: { type: boolean }
...
    PendingApprovalItem:
      type: object
      description: One pending document; includes approval_trails (history of actions, newest first by created_at).
      properties:
        id: { type: integer }
        category: { type: string }
        type: { type: string, description: "e.g. SpecialMemo, Matrix, Activity, OtherMemo" }
        status: { type: string }
        title: { type: string }
        division: { type: string }
        submitted_by: { type: string }
        date_received: { type: string, nullable: true }
        view_url: { type: string }
        approval_level: { type: integer }
        workflow_role: { type: string, nullable: true }
        item_id: { type: integer }
        item_type: { type: string }
        approval_trails:
          type: array
          items: { $ref: '#/components/schemas/ApprovalTrailItem' }
          description: Full approval history; sorted by created_at descending (latest first).
...
security:
  - bearerAuth: []